// src/types.ts
export interface NatsMessage {
	/**
	 * Defines the message's subject.
	 * @type {string}
	 */
	subject: string;

	/**
	 * The message itself.
	 * @type {string}
	 */
	message: Uint8Array;

	/**
	 * Optional set of key/value associated with the message.
	 * @type {NatsMessageHeaders | undefined}
	 */
	headers?: NatsMessageHeaders;
};

/**
 * Contents of a presistnt message.
 * @interface NatsStreamMessage
 */
export interface NatsStreamMessage extends NatsMessage {
	/**
	 * The sequence number of the message in the stream.
	 * @type {number}
	 */
	sequence: number;

	/**
	 * The time the message was received.
	 * @type {Date}
	 */
	timestamp: Date;
};

/**
 * A persistent message that must be acknowledged.
 * @interface NatsStreamMessageAck
 */
export interface NatsStreamMessageAck extends NatsStreamMessage {
	/**
	 * Call this method to acknowledge reception and handling of this message.
	 * @method ack
	 */
	ack: () => void

	/**
	 * Indicate to the JetStream server that processing of the message failed,
	 * and that it should be resent after the spefied number of milliseconds.
	 * @param {number|undefined} millis
	 */
	nak: (millis?: number) => void
};

/**
 * A set of key/value items that will be associated to a given message.
 * @interface NatsMessageHeaders
 **/
export interface NatsMessageHeaders {
	[key: string]: string;
};
// ...
export interface NatsConsumerOptions {
	/**
	 * The topic to subscribe. If not specified, the stream's filter will be used.
	 * @type {string|undefined}
	 */
	subjectFilter?: string;

	/**
	 * Indicates the start point to use to deliver messages.
	 *   All: Start receiving from the earliest available message.
	 *   New: Start receiving messages that were created after the consumer was created.
	 *   Last: Start receiving messages with the last message added to the stream, or the
	 *         last message in the stream that matches the consumer's filter subject if defined.
	 *   Subject-Last: Start with the latest one for each filtered subject currently in the stream.
	 *   Sequence: Start at the first message having the sequence number or the next one available.
	 *   Time: Start with messages on or after this time.
	 * @type {'all' | 'new' | 'last' | 'sequence' | 'time' | 'subject-last'}
	 * @default new
	 */
	deliverPolicy?: 'all' | 'new' | 'last' | 'sequence' | 'time' | 'subject-last';

	/**
	 * The sequence from which to start delivery messages.
	 * @type {number|undefined}
	 */
	deliverStartSequence?: number;

	/**
	 * The date time from which to start delivering messages
	 * @type {Date|undefined}
	 */
	deliverStartTime?: Date;

	/**
	 * Action to execute if the stream already exists. Can be: update, keep original or fail.
	 * @type {'update'|'keep'|'fail'|undefined}
	 * @default update
	 */
	existingAction?: 'update' | 'keep' | 'fail';
};
// ...
export interface NatsClientStreamInfo {
	/**
	 * Number of messages stored in the Stream
	 */
	messages: number;

	/**
	 * Combined size of all messages in the Stream
	 */
	bytes: number;

	/**
	 * Sequence number of the first message in the Stream
	 */
	firstSequence: number;

	/**
	 * The timestamp of the first message in the Stream
	 */
	firstTimestamp: Date;

	/**
	 * Sequence number of the last message in the Stream
	 */
	lastSequence: number;

	/**
	 * The timestamp of the last message in the Stream
	 */
	lastTimestamp: Date;
};

/**
 * Warps a JetStream stream.
 * @interface NatsClientStream
 */
export interface NatsClientStream {
	/**
	 * Gets stream information.
	 * @async
	 * @method info
	 */
	info: () => Promise<NatsClientStreamInfo>;

	/**
	 * Instructs the server to delete this stream.
	 * @async
	 * @method delete
	 */
	delete: () => Promise<void>;

	/**
	 * Gets an existing JetStream stream's message.
	 * @async
	 * @method getMessage
	 * @param {number} sequence - Sequence number of the message to obtain.
	 */
	getMessage: (sequence: number) => Promise<NatsStreamMessage>;

	/**
	 * Deletes an existing JetStream stream's message.
	 * @async
	 * @method deleteMessage
	 * @param {number} sequence - Message sequence number to delete.
	 */
	deleteMessage: (sequence: number) => Promise<void>;

	/**
	 * Gets an existing JetStream consumer.
	 * @async
	 * @method getConsumer
	 * @param {string} name - Name of the consumer.
	 */
	getConsumer: (name: string) => Promise<NatsClientConsumer>;

	/**
	 * Creates a new JetStream consumer.
	 * @async
	 * @method createConsumer
	 * @param {string} name - Name of the consumer.
	 * @param {NatsClientConsumer} opts - Consumer configuration options.
	 */
	createConsumer: (name: string, opts: NatsConsumerOptions) => Promise<NatsClientConsumer>;

	/**
	 * Deletes an existing JetStream consumer.
	 * It does not throw errors if the consumer does not exists.
	 * @async
	 * @method deleteConsumer
	 * @param {string} name - Name of the consumer to delete.
	 */
	deleteConsumer: (name: string) => Promise<void>;
};

/**
 * Warps a JetStream stream consumer.
 * @interface NatsClientConsumer
 */
export interface NatsClientConsumer {
	/**
	 * Instructs the server to delete this consumer.
	 * @async
	 * @method delete
	 */
	delete: () => Promise<void>;

	/**
	 * Starts a subscription for messages based on the given stream and consumer.
	 * @async
	 * @method subscribe
	 * @param {NatsStreamMessageCallback} cb - Asynchronous callback to call when a new message arrives.
	 */
	subscribe: (cb: NatsStreamMessageCallback) => Promise<void>;

	/**
	 * Ends the subscription.
	 * It does not throw errors if the connection is closed or the subscription does not exists.
	 * @async
	 * @method subscribe
	 */
	unsubscribe: () => Promise<void>;
};
// ...
export type NatsStreamMessageCallback = (msg: NatsStreamMessageAck) => Promise<boolean | undefined>;